Add sphere
This is a brute force algorithm, so could be optimized later.

The At() option is added to Box to allow us to specify a single
voxel box at a given location.

Note that the option to set the surface type of sphere is
WithSphereSurface() because WithSurface is already defined
for the BoxOption type.